requirements.txt
pandas
openai
chromadb
pytest
requests
python-dotenv
boto3